moved augurpy functions around and adapted docs
