--- docs/usage/cli.rst
@@ -1,6 +1,0 @@
-Command-line interface
-======================
-
-.. click:: pertpy.__main__:pertpy_cli
-   :prog: pertpy
-   :nested: full
