--- docs/usage/data/pertpy.api.data.burczynski_crohn.rst
@@ -1,12 +1,0 @@
-﻿:github_url: pertpy.api.data.burczynski_crohn
-
-pertpy.api.data.burczynski\_crohn
-=================================
-
-.. autofunction:: pertpy.api.data.burczynski_crohn
-
-.. _sphx_glr_backref_pertpy.api.data.burczynski_crohn:
-
-.. minigallery:: pertpy.api.data.burczynski_crohn
-   :add-heading: Gallery
-   :heading-level: -
