+++ docs/usage/data/pertpy.data.burczynski_crohn.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.data.burczynski_crohn
+
+pertpy.data.burczynski\_crohn
+=============================
+
+.. autofunction:: pertpy.data.burczynski_crohn
+
+.. _sphx_glr_backref_pertpy.data.burczynski_crohn:
+
+.. minigallery:: pertpy.data.burczynski_crohn
+   :add-heading: Gallery
+   :heading-level: -
