+++ docs/usage/differential_prioritization/pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization
+
+pertpy.tools.augurpy.differential\_prioritization.predict\_differential\_prioritization
+=======================================================================================
+
+.. autofunction:: pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization
+
+.. _sphx_glr_backref_pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization:
+
+.. minigallery:: pertpy.tools.augurpy.differential_prioritization.predict_differential_prioritization
+   :add-heading: Gallery
+   :heading-level: -
