+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.max_depth.rst
@@ -1,0 +1,8 @@
+:github_url: pertpy.tools.augurpy.estimator.Params.max_depth
+
+pertpy.tools.augurpy.estimator.Params.max\_depth
+================================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoattribute:: Params.max_depth
