+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.max_features.rst
@@ -1,0 +1,8 @@
+:github_url: pertpy.tools.augurpy.estimator.Params.max_features
+
+pertpy.tools.augurpy.estimator.Params.max\_features
+===================================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoattribute:: Params.max_features
