+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.n_estimators.rst
@@ -1,0 +1,8 @@
+:github_url: pertpy.tools.augurpy.estimator.Params.n_estimators
+
+pertpy.tools.augurpy.estimator.Params.n\_estimators
+===================================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoattribute:: Params.n_estimators
