+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.penalty.rst
@@ -1,0 +1,8 @@
+:github_url: pertpy.tools.augurpy.estimator.Params.penalty
+
+pertpy.tools.augurpy.estimator.Params.penalty
+=============================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoattribute:: Params.penalty
