+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.random_state.rst
@@ -1,0 +1,8 @@
+:github_url: pertpy.tools.augurpy.estimator.Params.random_state
+
+pertpy.tools.augurpy.estimator.Params.random\_state
+===================================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoattribute:: Params.random_state
