+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.Params.rst
@@ -1,0 +1,39 @@
+﻿:github_url: pertpy.tools.augurpy.estimator.Params
+
+pertpy.tools.augurpy.estimator.Params
+=====================================
+
+.. currentmodule:: pertpy.tools.augurpy.estimator
+
+.. autoclass:: Params
+
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+      :toctree: .
+   
+      ~pertpy.tools.augurpy.estimator.Params.max_depth
+      ~pertpy.tools.augurpy.estimator.Params.max_features
+      ~pertpy.tools.augurpy.estimator.Params.n_estimators
+      ~pertpy.tools.augurpy.estimator.Params.penalty
+      ~pertpy.tools.augurpy.estimator.Params.random_state
+   
+   
+
+   
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+      :toctree: .
+   
+   
+   
+
+   .. _sphx_glr_backref_pertpy.tools.augurpy.estimator.Params:
+
+   .. minigallery:: pertpy.tools.augurpy.estimator.Params
+       :add-heading: Gallery
+       :heading-level: -
