+++ docs/usage/estimator/pertpy.tools.augurpy.estimator.create_estimator.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.tools.augurpy.estimator.create_estimator
+
+pertpy.tools.augurpy.estimator.create\_estimator
+================================================
+
+.. autofunction:: pertpy.tools.augurpy.estimator.create_estimator
+
+.. _sphx_glr_backref_pertpy.tools.augurpy.estimator.create_estimator:
+
+.. minigallery:: pertpy.tools.augurpy.estimator.create_estimator
+   :add-heading: Gallery
+   :heading-level: -
