+++ docs/usage/evaluate/pertpy.tools.augurpy.evaluate.predict.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.tools.augurpy.evaluate.predict
+
+pertpy.tools.augurpy.evaluate.predict
+=====================================
+
+.. autofunction:: pertpy.tools.augurpy.evaluate.predict
+
+.. _sphx_glr_backref_pertpy.tools.augurpy.evaluate.predict:
+
+.. minigallery:: pertpy.tools.augurpy.evaluate.predict
+   :add-heading: Gallery
+   :heading-level: -
