+++ docs/usage/plot/pertpy.plot.augurpy.differential_prioritization.dp_scatter.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.plot.augurpy.differential_prioritization.dp_scatter
+
+pertpy.plot.augurpy.differential\_prioritization.dp\_scatter
+============================================================
+
+.. autofunction:: pertpy.plot.augurpy.differential_prioritization.dp_scatter
+
+.. _sphx_glr_backref_pertpy.plot.augurpy.differential_prioritization.dp_scatter:
+
+.. minigallery:: pertpy.plot.augurpy.differential_prioritization.dp_scatter
+   :add-heading: Gallery
+   :heading-level: -
