+++ docs/usage/plot/pertpy.plot.augurpy.important_features.important_features.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.plot.augurpy.important_features.important_features
+
+pertpy.plot.augurpy.important\_features.important\_features
+===========================================================
+
+.. autofunction:: pertpy.plot.augurpy.important_features.important_features
+
+.. _sphx_glr_backref_pertpy.plot.augurpy.important_features.important_features:
+
+.. minigallery:: pertpy.plot.augurpy.important_features.important_features
+   :add-heading: Gallery
+   :heading-level: -
