+++ docs/usage/plot/pertpy.plot.augurpy.lollipop.lollipop.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.plot.augurpy.lollipop.lollipop
+
+pertpy.plot.augurpy.lollipop.lollipop
+=====================================
+
+.. autofunction:: pertpy.plot.augurpy.lollipop.lollipop
+
+.. _sphx_glr_backref_pertpy.plot.augurpy.lollipop.lollipop:
+
+.. minigallery:: pertpy.plot.augurpy.lollipop.lollipop
+   :add-heading: Gallery
+   :heading-level: -
