+++ docs/usage/plot/pertpy.plot.augurpy.scatterplot.scatterplot.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.plot.augurpy.scatterplot.scatterplot
+
+pertpy.plot.augurpy.scatterplot.scatterplot
+===========================================
+
+.. autofunction:: pertpy.plot.augurpy.scatterplot.scatterplot
+
+.. _sphx_glr_backref_pertpy.plot.augurpy.scatterplot.scatterplot:
+
+.. minigallery:: pertpy.plot.augurpy.scatterplot.scatterplot
+   :add-heading: Gallery
+   :heading-level: -
