+++ docs/usage/read_load/pertpy.tools.augurpy.read_load.load.rst
@@ -1,0 +1,12 @@
+﻿:github_url: pertpy.tools.augurpy.read_load.load
+
+pertpy.tools.augurpy.read\_load.load
+====================================
+
+.. autofunction:: pertpy.tools.augurpy.read_load.load
+
+.. _sphx_glr_backref_pertpy.tools.augurpy.read_load.load:
+
+.. minigallery:: pertpy.tools.augurpy.read_load.load
+   :add-heading: Gallery
+   :heading-level: -
