--- pertpy/tl/__init__.py
@@ -1,1 +1,0 @@
-from pertpy.tl._example import add_me
