--- pertpy/tl/_example.py
@@ -1,2 +1,0 @@
-def add_me(a: int, b: int) -> int:
-    return a + b
