--- tests/tl/test_example.py
@@ -1,5 +1,0 @@
-from pertpy.tl._example import add_me
-
-
-def test_add_me():
-    assert add_me(39, 3) == 42
